+++ boot.js
@@ -1,7 +1,11 @@
 const {ValueResolvingConfig,EphemeralConfig,ConfigFactory} = require ('@alt-javascript/config');
-const {LoggerRegistry,LoggerFactory,ConfigurableLogger} = require('@alt-javascript/logger');
+const {LoggerCategoryCache,LoggerFactory,ConfigurableLogger} = require('@alt-javascript/logger');
 
-module.exports = function (configArg,loggerFactoryArg,loggerRegistryArg) {
+module.exports = function (context) {
+
+    let configArg = context?.config;
+    let loggerFactoryArg = context?.loggerFactory;
+    let loggerCategoryCacheArg = context?.loggerFactory;
 
     let browser = !(typeof window == 'undefined');
 
@@ -27,21 +31,21 @@ module.exports = function (configArg,loggerFactoryArg,loggerRegistryArg) {
         throw new Error ('Unable to detect config, is \'config\' declared or provided?');
     }
 
-    let _loggerRegistry = null;
-    if (!(typeof loggerRegistry == 'undefined')){
-        _loggerRegistry = loggerRegistry;
+    let _loggerCategoryCache = null;
+    if (!(typeof loggerCategoryCacheArg == 'undefined')){
+        _loggerCategoryCache = loggerCategoryCacheArg;
     }
-    if (browser && window?.loggerRegistry){
-        _config = window.loggerRegistry;
+    if (browser && window?.loggerCategoryCache){
+        _config = window.loggerCategoryCache;
     }
 
-    _loggerRegistry = _loggerRegistry || loggerRegistryArg || new LoggerRegistry();
+    _loggerCategoryCache = _loggerCategoryCache || loggerCategoryCacheArg || new LoggerCategoryCache();
 
     let _loggerFactory = null;
     if (!(typeof loggerFactory == 'undefined')){
         _loggerFactory = loggerFactory;
     }
-    _loggerFactory = _loggerFactory || loggerFactoryArg || new LoggerFactory(_config,_loggerRegistry,ConfigurableLogger.DEFAULT_CONFIG_PATH);
+    _loggerFactory = _loggerFactory || loggerFactoryArg || new LoggerFactory(_config,_loggerCategoryCache,ConfigurableLogger.DEFAULT_CONFIG_PATH);
 
     let _globalref = null;
     if (browser){
@@ -50,7 +54,7 @@ module.exports = function (configArg,loggerFactoryArg,loggerRegistryArg) {
         _globalref = global;
     }
     _globalref.boot = {contexts: {root:{config:_config}}};
-    _globalref.boot.contexts.root.loggerRegistry = _loggerRegistry;
+    _globalref.boot.contexts.root.loggerCategoryCache = _loggerCategoryCache;
     _globalref.boot.contexts.root.loggerFactory = _loggerFactory;
 
 }